Full specification genearted.
This commit is the first commit that appears to have a full
specification. This includes structural-hashes for the specification,
structural-hashes for each type, and tagging of partials by a
structural-hash.

Still needs tests, needs some cleanup, and it would be really good to
get a total interface to the TypeIdMap so that I don't have to litter a
bunch of fromJust's all over. I realize they are safe the way they are
constructed, but it would be nice to isolate that safe-unsafety to one
location.